fixing folder for priv + public key flag

+++ main_test.go
@@ -1,7 +1,9 @@
 package main
 
 import (
+	"fmt"
 	"os"
+	"os/exec"
 	"path"
 	"testing"
 
@@ -10,14 +12,9 @@ import (
 
 func TestKeyGen(t *testing.T) {
 
-	os.Args = []string{"drand", "keygen"}
-	main()
 	fs := DefaultFileStore()
 	defer os.RemoveAll(fs.KeyFile)
 	defer os.RemoveAll(fs.PublicFile)
-	// no address
-	_, err := fs.LoadKey()
-	require.Error(t, err)
 
 	// valid address
 	os.Args = []string{"drand", "keygen", "127.0.0.1:8080"}
@@ -27,15 +24,39 @@ func TestKeyGen(t *testing.T) {
 	require.NotNil(t, priv.Public)
 
 	// custom file
-	tmp := os.TempDir()
-	fullPath := path.Join(tmp, defaultKeyFile+privateExtension)
-	fs.KeyFile = fullPath
-	fs.PublicFile = publicFile(fullPath)
-	defer os.RemoveAll(fs.KeyFile)
-	defer os.RemoveAll(fs.PublicFile)
-	os.Args = []string{"drand", "keygen", "--" + keyFileFlagName, fullPath, "127.0.0.1:8080"}
+	tmp := path.Join(os.TempDir(), "drand")
+	os.MkdirAll(tmp, 0777)
+	fs.KeyFile = path.Join(tmp, defaultKeyFile+privateExtension)
+	fs.PublicFile = publicFile(fs.KeyFile)
+	//defer os.RemoveAll(fs.KeyFile)
+	//defer os.RemoveAll(fs.PublicFile)
+	os.Args = []string{"drand", "keygen", "--" + keyFolderFlagName, tmp, "127.0.0.1:8080"}
 	main()
 	priv, err = fs.LoadKey()
 	require.Nil(t, err)
 	require.NotNil(t, priv.Public)
+}
+
+// https://stackoverflow.com/questions/26225513/how-to-test-os-exit-scenarios-in-go
+func TestKeyGenInvalid(t *testing.T) {
+	varEnv := "CRASHCRASH"
+	if os.Getenv(varEnv) == "1" {
+		os.Args = []string{"drand", "keygen"}
+		fmt.Println("bri")
+		main()
+		return
+	}
+	cmd := exec.Command(os.Args[0], "-test.run=TestKeyGenInvalid")
+	cmd.Env = append(os.Environ(), varEnv+"=1")
+	err := cmd.Run()
+	if e, ok := err.(*exec.ExitError); ok && e.Success() {
+		t.Fatalf("KeyGenInvalid should have failed")
+	}
+	fs := DefaultFileStore()
+	defer os.RemoveAll(fs.KeyFile)
+	defer os.RemoveAll(fs.PublicFile)
+	// no address
+	_, err = fs.LoadKey()
+	require.Error(t, err)
+
 }
